Precompilation kills runtime performance because of Julia bugs.
Master:
```julia
julia> @benchmark matmul!($Cf64, $Af64, $Bf64)
BenchmarkTools.Trial: 3727 samples with 1 evaluation.
 Range (min … max):  258.678 μs … 279.208 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     265.012 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   265.171 μs ±   1.905 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

                     ▁▂▄▄▆█▆▆▇▇▇█▇▇▃▄▃▁▁                         
  ▂▁▁▁▁▂▂▁▂▃▂▃▃▃▄▄▅▅▇████████████████████▆▆▅▄▅▄▄▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂ ▅
  259 μs           Histogram: frequency by time          272 μs <

 Memory estimate: 272 bytes, allocs estimate: 17.
```
PR:
```julia
julia> @benchmark matmul!($Cf64, $Af64, $Bf64)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  23.756 μs …  43.892 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     24.462 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.547 μs ± 586.526 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

           ▁▃▅▆██▆▄▃▁                                           
  ▂▂▂▂▂▃▃▅▆██████████▇▅▄▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▁▁▂▁▁▁▁▁▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  23.8 μs         Histogram: frequency by time         26.7 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
@ranocha have you seen this?

This was the benchmark script I used:
```julia
julia> A = rand(-1_000:1_000, 200, 200);

julia> B = rand(-1_000:1_000, 200, 200);

julia> C = similar(A);

julia> @benchmark mul!($C, $A, $B)
BenchmarkTools.Trial: 411 samples with 1 evaluation.
 Range (min … max):  2.283 ms …  12.646 ms  ┊ GC (min … max): 0.00% … 77.02%
 Time  (median):     2.372 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.430 ms ± 515.259 μs  ┊ GC (mean ± σ):  0.98% ±  3.80%

         ██        ▁                                           
  ▅▂▁▁▁▂▃██▄▄▅▅▅▄▄▇█▅▃▄▆▄▃▂▃▁▂▁▂▁▁▂▁▂▃▁▁▁▃▁▂▂▂▃▃▃▃▂▁▂▂▁▄▅▃▁▃▂ ▃
  2.28 ms         Histogram: frequency by time        2.67 ms <

 Memory estimate: 30.77 KiB, allocs estimate: 4.

julia> @benchmark matmul!($C, $A, $B)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  90.987 μs … 118.231 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     91.633 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   92.750 μs ±   2.644 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃▇█▅    ▂▄▃▁ ▃▆▅▁     ▁▁                                     ▂
  ████▇▃▄▆████▇████▁▃▁▆███▇▆▅▅▁▁▃▁▁▁▁▃▁▁▁▁▃▁▁▁▁▁▁▁▅▆▆▅▄▆▅▆▆▇▆▄ █
  91 μs         Histogram: log(frequency) by time       106 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> 2.369e-3 / 91.302e-6
25.946857681102276

julia> Af64 = Float64.(A); Bf64 = Float64.(B); Cf64 = similar(Af64);

julia> @benchmark mul!($Cf64, $Af64, $Bf64)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  14.808 μs … 34.542 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     15.224 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.367 μs ±  1.002 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

     ▄█▆▂                                                      
  ▂▃▆████▆▄▃▂▂▂▁▁▂▂▂▂▂▁▂▁▁▁▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▁▂▂▂▂ ▃
  14.8 μs         Histogram: frequency by time        19.7 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark matmul!($Cf64, $Af64, $Bf64)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  23.756 μs …  43.892 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     24.462 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.547 μs ± 586.526 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

           ▁▃▅▆██▆▄▃▁                                           
  ▂▂▂▂▂▃▃▅▆██████████▇▅▄▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▁▁▂▁▁▁▁▁▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  23.8 μs         Histogram: frequency by time         26.7 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> Cf64 == C
true
```