@ranocha 2023-06-28
> @ranocha have you seen this?

No - but I have to admit that I haven't been running a lot of `matmul!`-based code recently, focused on other parts of our codebase instead. 

Thanks a lot for fixing this! Do you have an idea which Julia bugs are involved here?