@chriselrod 2023-06-28
> Thanks a lot for fixing this! Do you have an idea which Julia bugs are involved here?

No, but bugs such as this have been known for a long time.
TriangularSolve.jl had to disable precompilation as well for this reason.